Final
I ran into a lot of issues twords the end, I had the server the apis and almost anything ready but 4 I couldnt figure out these small errors. I found another method online which led to me doing that using only google api to retrieve the calendar data and not a server, then as I continued to do that I saw that the errors were because using a server to obtain the data was not the right way to do it because google helped lay out the map on how to do it using their apis.